Add test.txt
Practice writing commit messages. This file will be deleted later.

In fact, it will be deleted on the next commit. Inside the file exists nothing.

As the file name is test, this commit is just like a test.

Lets try bullet points:

1. I mean number.

2. Number 2.

3. You know what crazy I almost hit enter.

IM NOT FINISHED YET

- It is capital test

I can't wait() to see it!